refactor: remove cyclic authz dependencies
