+++ x/authz/keeper/keeper.go
@@ -134,11 +134,15 @@ func (k Keeper) DispatchActions(ctx context.Context, grantee sdk.AccAddress, msg
 			if err != nil {
 				return nil, err
 			}
+			updated, ok := resp.Updated.(authz.Authorization)
+			if !ok {
+				return nil, fmt.Errorf("expected %T but got %T", (authz.Authorization)(nil), resp.Updated)
+			}
 
 			if resp.Delete {
 				err = k.DeleteGrant(ctx, grantee, granter, sdk.MsgTypeURL(msg))
 			} else if resp.Updated != nil {
-				err = k.update(ctx, grantee, granter, resp.Updated)
+				err = k.update(ctx, grantee, granter, updated)
 			}
 			if err != nil {
 				return nil, err
